requirements.txt
certifi==2019.6.16
chardet==3.0.4
docker==4.0.2
idna==2.8
pyaml==19.4.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
